requirements.txt
feedparser
wordpress-xmlrpc
openai
